postgres_to_es/requirements.txt
psycopg2-binary==2.9.1
pydantic==1.8.2
elasticsearch==7.14.2
backoff==1.11.1
redis==3.5.3
python-dotenv==0.19.0